Separate methods from main.rb file
